Add docs for doctypes, ES_TIMEOUT, and ES building
ElasticUtils builds queries with a series of assumptions about how the
index is laid out.  This documents that.

Adds documentation for ES_TIMEOUT.

Added a note about creating ES objects if you want an ES that isn't
cached thread-local.  This makes it easier to build an ES with
settings different than those in settings.py.  It's just a note to
make things easier.