Overhaul docs
This adds contributor docs and cleans up some of the elasticutils
usage docs, too. It adds a lot of examples and expands on the existing
text.

This also adds a requirements-dev.txt which makes it a little easier
to install for hacking.

Still a lot of docs-work to do, but I think this is a nice first pass.