More query doc work
Ricky asked what the elasticsearch JSON looks like if you have filters
and queries in it. I thought about that and decided to add more to the
overview section so that there's a big example coupled with what it
equates to in curl and then the rest of the document talks about
individual pieces.

I also added a separate section about S.