Tweak F so that it allows for F()
This is something I implemented in the sumo branch to allow for an empty
F. It makes it a lot easier to start with an F and add filters depending
on what the user requested and also handle the case where there are
no filters.

e.g.

    f = F()
    f = f | F(tag='awesome')

Prior to this change, that would result in:

    {"or_": [{}, {"tag": "awesome"}]}

which makes ES grumpy and for good reason.