Fix list results
If there was only one item in self.fields, then you'd end up with
a flat list of things instead of a list of tuples. We want that
to be consistent, though, so if we know we're in that situation, we
fix it accordingly.

This is a little awkward if you really want a flat list. Maybe we
should make that easier in the future.