Issue 43. Reimplement facets
This reworks facets so that the API is cleaner and easier to use. It
now has a .facet() which acts as a shorthand but doesn't do things
you might not expect and it has a .facet_raw() which yields unto
you the full expressive power of the ES facet.

This adds some more tests and documentation and updates the CHANGELOG,
too.