Implement execute
Previously you could only force an S to execute the search as a
side effect of requesting something out of it (count, facet_counts,
an iterable of results). This allows you to execute the search
and get back a SearchResults instance which has some interesting
things on it like a .results property which has the complete
ES response body.