Clean up django tests; make django requirement for dev
This nixes all the code that allowed you to run tests without having
django installed. Now, in order to run the tests, you need to have
it installed--it's a requirement for developing elasticutils but
not using elasticutils.

This also breaks up test_django.py into a bunch of separate files
and moves the infrastructure to __init__.py.