Fix test failures
A bunch of tests would occasionally fail because the filter and query
portions of the Elasticsearch query are generated from dicts which
get turned into lists and thus we don't know what order they come in.
Thus two equivalent lists could be in different orders, but we consider
them "the same" because we don't really care about the order. Hence our
new compare function: "eqish_".

The other way to solve this is to use ordered dicts which keep things in
the order you add them. At some point, it might make sense to change it
but for now I'll just deal with this in the tests.