Move results metadata to es_meta object
Django templates don't allow you to expose variables that start with an
_. So having:

    obj._highlight

sucked.

We've got a fair number of these metadata items and I'm concerned with
each new one we add, we increase the likelihood that there's a naming
conflict with actual data.

So this moves all those items into an es_meta object which you can
access like this:

    obj.es_meta.highlight

The one exception is _id--you can now access the id Elasticsearch gives
the document with both:

    obj._id

and:

    obj.es_meta.id

Fixes #205