Add elasticutils.estestcase.ESTestCase
This makes the ESTestCase "public". Previously, it was intertwined with
our test suite.

Further, it adds documentation for it to the api.

In the process of doing that, I nixed a bunch of skip_tests stuff. If
someone doesn't want it to run tests, they should wrap the class on
their own. That's a lot easier than us providing scaffolding that is
either generic (hard) or requires a specific test framework
(pain-in-the-ass-for-users).

Fixes #215